refactor(cli): decompose run.rs into focused submodules [OPE-378]
Split the monolithic 775-line run.rs into gateway, pairing, lifecycle,
and test modules with clear separation of concerns.
